+++ CHANGES.rst
@@ -6,6 +6,11 @@ Products.MeetingPROVHainaut Changelog
 
 - Added `testVotes.py` as it is launched now by `Products.MeetingCommunes`.
   [gbastien]
+- Added `test_CompletenessEvaluationAskedAgain` that shows that completeness
+  evaluation is asked correctly (test fixes in
+  `Products.MeetingCommunes.adapters._will_ask_completeness_eval_again` and
+  `Products.MeetingCommunes.adapters._doWaitAdvices`).
+  [gbastien]
 
 4.2b3 (2020-10-14)
 ------------------